Moved ticket status definitions to dedicated sources
